separate default renderer from renderer map
